refactor: WinnerLotto 제거 -> Lotto, Utils 제거 -> Numbers 에게 위임
